+++ src/core/RemoteFiles.h
@@ -73,7 +73,7 @@ public:
   const TRemoteToken * Find(const std::wstring & Name) const;
   void Log(TTerminal * Terminal, const wchar_t * Title);
 
-  int GetCount() const;
+  size_t GetCount() const;
   const TRemoteToken * GetToken(int Index) const;
 
 private:
@@ -531,7 +531,7 @@ bool IsUnixHiddenFile(const std::wstring Path);
 std::wstring AbsolutePath(const std::wstring & Base, const std::wstring & Path);
 std::wstring FromUnixPath(const std::wstring Path);
 std::wstring ToUnixPath(const std::wstring Path);
-std::wstring MinimizeName(const std::wstring FileName, int MaxLen, bool Unix);
+std::wstring MinimizeName(const std::wstring FileName, size_t MaxLen, bool Unix);
 std::wstring MakeFileList(TStrings * FileList);
 TDateTime ReduceDateTimePrecision(TDateTime DateTime,
   TModificationFmt Precision);
